Optimize peers usage
